By default go back an extra 10 seconds (overlapping with previous activity requests) This means we get repeat activities, so keep track of activities, and don't process it if it has been processed before.
This is because, in particular I am seeing tasks being edited which are unknown, I suspect the task create activity is not being received somehow ...
If it takes pivotal tracker a bit of time to record an activity, and the activity gets recorded at the time of activity not the time of recording,
then we could miss some. So hopefully it doesn't take pivotal tracker more ten 10 seconds from activity to storage.